Introduce sniff to detect issues with minimum_wp_version in phpcs.xml
The WP Coding Standards using minimum_wp_version to configure many of their sniffs. The version is validated against `float` so if a patch version is included we always get the fallback version.

We can't fix the logic directly, so instead we detect issues with the configuration and throw an error, so it will be fixed.