Instead of waiting for php `phpcompatibility/php-compatibility` to release use develop
It has been 2 years and still no version 10.
https://github.com/PHPCompatibility/PHPCompatibility/issues/1343

Instead of ignoring the rule which prevents true detecting, use the develop version with a version 9 alias.

The alias is required to use with phpcompatibility-wp which expects version ^9.

Once version 10 is released and phpcompatibility-wp supports it, remove the entry from require-dev.

When scanning Enums
 "$this" can no longer be used in a plain function or method since PHP 7.1. (PHPCompatibility.Variables.ForbiddenThisUseContexts.OutsideObjectContext)